Add condition to beautify htmls with only one < pre > node
The algorithmus was applied to any url with .js at the ending.
The led to an inteference at pages that aren't (at least not only)
js codes, like github pages from files that end with .js.
This condition filters the urls, applying the beautification
only to htmls that have one 'pre' tag and nothing else.

Resolves: #3